Refactor UI files and classes
